Move RTH recovery test (last non-async test) to Async file
